+++ src/context/theme/index.ts
@@ -1,4 +1,3 @@
 export * from './ThemeContext';
-export * from './ThemeController';
 export * from './useTheme';
 export * from './withTheme';
